Split runc shim into plugin components
